Rescaneo de la base de datos por errores presentados

+++ NeoAPI/Models/Neo/LibroNove.cs
@@ -13,7 +13,7 @@ public partial class LibroNove
 
     public string Lndiscrepa { get; set; } = null!;
 
-    public int LntiePerMi { get; set; }
+    public double LntiePerMi { get; set; }
 
     public string LnfichaRes { get; set; } = null!;
